Remove JS implementation
