--- src/serialization-schema.fbs
@@ -1,21 +1,0 @@
-namespace Serialization;
-
-struct Point {
-  row:int;
-  column:int;
-}
-
-table Change {
-  oldStart:Point;
-  newStart:Point;
-  oldExtent:Point;
-  newExtent:Point;
-  newText:string;
-  oldText:string;
-}
-
-table Patch {
-  changes:[Change];
-}
-
-root_type Patch;
