+++ test/helpers/text-helpers.js
