+++ test/tests.cc
@@ -1,2 +1,5 @@
+// Because this file is slow to compile, we separate it from patch_test.cc
+// for a faster feedback loop
+
 #define CATCH_CONFIG_MAIN
 #include <catch.hpp>
